refactor(views): extract layout state + decompose ManuscriptView rendering
P3 of the architecture refactor.

ManuscriptView: split ManuscriptViewUI into ManuscriptMobileLayout
(header + full-bleed editor + drawers) and ManuscriptDesktopLayout
(3-pane navigator · editor+research · inspector). Shared UI state
(nav tab, mobile drawers, focus mode, resizable panels, research-split
Escape handler) moves into hooks/useManuscriptLayout.ts; the view is now
a thin orchestrator that renders both responsive layouts.

WriterViewUI: extract local state (mobile tab, panel collapse, focus
mode, flow-mode Escape handler, swipe-to-switch wiring) into
hooks/useWriterLayout.ts; rendering unchanged.

No behavior change — responsive show/hide, ARIA tablists, drawers, and
keyboard handlers are preserved. Verified: typecheck, lint (0 warnings),
38 tests pass (ManuscriptView, WriterView, useWriterView).
